Add files via upload
This graph displays the averaged number of possible moves at different stages(rounds) of the game. 
We can see that up until round number 4, 4 players have an equal(on average) number of possible moves.
The game starts leaning towards player1,2's favor starting from round number 5. 
With the 100 simulations we have, player 1,2 have a 30% winning chance, and player 3,4 have around 20 each. This is reflected in round number 5, as we can see that number of possible moves shows a clear gap between 1,2 vs 3,4.
Starting with round 10, this gap is even clearer. 
Yet, the fact that player3,4 have a combined winning chance of 40%s shows that at later stages of the game, the strategy of always moving the biggest block does not guarantee a win. In other words, this strategy is useless at later stages of the game, which makes sense, because everyone would have very small pieces, and either placing a size 3 block or a size 2 block does not make a big difference at this stage. 
What we can possibly conclude is that, at late stages of the game, those who place pieces first have a bigger winning chance.